Code Verbessert und Anmeldung Und Registrierung soweit fertig
