+++ core/src/Connection/Client.java
@@ -1,15 +1,12 @@
 package Connection;
 
-import Screens.StartScreen;
 import io.socket.client.IO;
 import io.socket.client.Socket;
-import io.socket.emitter.Emitter;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.net.URISyntaxException;
 import java.util.Arrays;
-import java.util.Objects;
 
 
 public class Client {
